remove preprocessing, add features, move functions to metadata
