+++ .gitignore
@@ -6,3 +6,4 @@
 **/*.wav
 **/*.mp3
 **/*.json
+**/*.csv
